Add ConnectivityStore, BonjourDiscoveryService, DiscoveryViewModel, and supporting files for Phase 1 & 2
