+++ SprinklerMobile/Services/BonjourDiscoveryService.swift
@@ -17,38 +17,24 @@ final class CurrentValueSubject<Output, Failure: Error> {
 }
 #endif
 
-struct DiscoveredDevice: Identifiable, Equatable {
-    let id: String         // "\(host ?? name):\(port)"
-    let name: String
-    let host: String?
-    let ip: String?
-    let port: Int
-
-    var baseURLString: String {
-        if let h = host { return "http://\(h):\(port)" }
-        if let ip = ip {
-            // bracket IPv6
-            if ip.contains(":") { return "http://[\(ip)]:\(port)" }
-            return "http://\(ip):\(port)"
-        }
-        return ""
-    }
-}
-
+/// Protocol describing the discovery interface used by the app.
 protocol BonjourDiscoveryProviding: AnyObject {
+    /// Publisher emitting currently discovered devices.
     var devicesPublisher: AnyPublisher<[DiscoveredDevice], Never> { get }
+    /// Starts Bonjour discovery.
     func start()
+    /// Stops Bonjour discovery.
     func stop()
+    /// Refreshes the discovery session.
     func refresh()
 }
 
-/// Build that **always compiles**. If Bonjour is unavailable or permission is denied,
-/// we still provide a no-op implementation so app builds & runs.
+/// Stubbed discovery service that keeps the project building even without Bonjour integration.
 final class BonjourDiscoveryService: BonjourDiscoveryProviding {
     private let subject = CurrentValueSubject<[DiscoveredDevice], Never>([])
     var devicesPublisher: AnyPublisher<[DiscoveredDevice], Never> { subject.eraseToAnyPublisher() }
 
-    func start() { /* real impl may go here; stub OK for build */ }
+    func start() { /* TODO: Implement NetServiceBrowser */ }
     func stop()  { /* noop */ }
     func refresh() { start() }
 }
