+++ SprinklerMobile/Services/HealthChecker.swift
@@ -3,6 +3,12 @@ import Foundation
 import FoundationNetworking
 #endif
 
+/// Abstraction used to verify sprinkler controller connectivity.
+protocol ConnectivityChecking {
+    func check(baseURL: URL) async -> ConnectivityState
+}
+
+/// Performs connectivity checks against the sprinkler controller's status endpoint.
 struct HealthChecker: ConnectivityChecking {
     private let session: URLSession
 
@@ -10,11 +16,16 @@ struct HealthChecker: ConnectivityChecking {
         self.session = session
     }
 
+    /// Invokes `/api/status` on the provided base URL and returns the resulting connectivity state.
     func check(baseURL: URL) async -> ConnectivityState {
         var statusURL = baseURL
         statusURL.append(path: "/api/status")
 
-        var req = URLRequest(url: statusURL, cachePolicy: .reloadIgnoringLocalCacheData, timeoutInterval: 8)
+        var req = URLRequest(
+            url: statusURL,
+            cachePolicy: .reloadIgnoringLocalCacheData,
+            timeoutInterval: 8
+        )
         req.httpMethod = "GET"
         req.addValue("application/json", forHTTPHeaderField: "Accept")
 
@@ -23,7 +34,6 @@ struct HealthChecker: ConnectivityChecking {
             guard let http = resp as? HTTPURLResponse, (200..<300).contains(http.statusCode) else {
                 return .offline(errorDescription: "Bad status")
             }
-            // valid JSON object = connected
             let obj = try JSONSerialization.jsonObject(with: data) as? [String: Any]
             return (obj != nil) ? .connected : .offline(errorDescription: "Non-JSON")
         } catch {
