+++ SprinklerMobile/Stores/ConnectivityStore.swift
@@ -12,16 +12,16 @@ struct Published<Value> {
 protocol ObservableObject {}
 #endif
 
-protocol ConnectivityChecking {
-    func check(baseURL: URL) async -> ConnectivityState
-}
-
+/// Stores the connectivity settings and state for communicating with the sprinkler controller.
 @MainActor
 final class ConnectivityStore: ObservableObject {
+    /// Base URL string for the sprinkler controller API. Persists to user defaults on change.
     @Published var baseURLString: String {
         didSet { defaults.set(baseURLString, forKey: Self.udKey) }
     }
+    /// Current connectivity state for the controller.
     @Published var state: ConnectivityState = .offline(errorDescription: nil)
+    /// Whether a connectivity check is currently in progress.
     @Published var isChecking: Bool = false
 
     private let checker: ConnectivityChecking
@@ -32,21 +32,18 @@ final class ConnectivityStore: ObservableObject {
     init(checker: ConnectivityChecking = HealthChecker(), defaults: UserDefaults = .standard) {
         self.defaults = defaults
         let saved = defaults.string(forKey: Self.udKey)
-        if let saved, !saved.isEmpty {
-            self.baseURLString = saved
-        } else {
-            self.baseURLString = Self.defaultBase
-            defaults.set(Self.defaultBase, forKey: Self.udKey)
-        }
+        self.baseURLString = (saved?.isEmpty == false) ? saved! : Self.defaultBase
         self.checker = checker
     }
 
+    /// Public entry point for refreshing connectivity state.
     func refresh() async { await testConnection() }
 
+    /// Tests connectivity to the configured base URL, updating the published state accordingly.
     func testConnection() async {
         let trimmed = baseURLString.trimmingCharacters(in: .whitespacesAndNewlines)
         guard let url = ConnectivityStore.normalizedBaseURL(from: trimmed) else {
-            state = .offline(errorDescription: "Invalid URL")
+            self.state = .offline(errorDescription: "Invalid URL")
             return
         }
         isChecking = true
@@ -55,6 +52,7 @@ final class ConnectivityStore: ObservableObject {
         await MainActor.run { self.state = result }
     }
 
+    /// Normalizes a human-entered string into a usable base URL, defaulting to HTTP when missing.
     static func normalizedBaseURL(from s: String) -> URL? {
         var str = s
         if !str.lowercased().hasPrefix("http://") && !str.lowercased().hasPrefix("https://") {
@@ -64,6 +62,7 @@ final class ConnectivityStore: ObservableObject {
     }
 }
 
+/// Represents connectivity state for the sprinkler controller.
 enum ConnectivityState: Equatable {
     case connected
     case offline(errorDescription: String?)
