+++ SprinklerMobile/ViewModels/DiscoveryViewModel.swift
@@ -1,78 +1,49 @@
 import Foundation
+import Dispatch
 #if canImport(Combine)
 import Combine
 #endif
 
-#if !canImport(SwiftUI)
-@propertyWrapper
-struct Published<Value> {
-    var wrappedValue: Value
-    init(wrappedValue: Value) { self.wrappedValue = wrappedValue }
-    var projectedValue: Published<Value> { self }
-}
-
-protocol ObservableObject {}
-#endif
-
-#if canImport(Combine)
+/// Bridges Bonjour discovery events into SwiftUI-friendly published properties.
 @MainActor
 final class DiscoveryViewModel: ObservableObject {
+    /// Discovered devices surfaced to the UI.
     @Published var devices: [DiscoveredDevice] = []
+    /// Indicates whether discovery is currently active.
     @Published var isBrowsing: Bool = false
+    /// Holds any user-facing error message from discovery.
     @Published var errorMessage: String?
 
     private let service: BonjourDiscoveryProviding
+    #if canImport(Combine)
     private var cancellables = Set<AnyCancellable>()
+    #endif
 
     init(service: BonjourDiscoveryProviding = BonjourDiscoveryService()) {
         self.service = service
+        #if canImport(Combine)
         service.devicesPublisher
             .receive(on: DispatchQueue.main)
             .sink { [weak self] in self?.devices = $0 }
             .store(in: &cancellables)
+        #endif
     }
 
+    /// Starts Bonjour discovery and flips the state flag.
     func start() {
         isBrowsing = true
         service.start()
     }
 
+    /// Requests a refresh from the service, typically re-browsing.
     func refresh() {
         isBrowsing = true
         service.refresh()
     }
 
+    /// Stops discovery and resets the browsing flag.
     func stop() {
         isBrowsing = false
         service.stop()
     }
 }
-#else
-@MainActor
-final class DiscoveryViewModel: ObservableObject {
-    @Published var devices: [DiscoveredDevice] = []
-    @Published var isBrowsing: Bool = false
-    @Published var errorMessage: String?
-
-    private let service: BonjourDiscoveryProviding
-
-    init(service: BonjourDiscoveryProviding = BonjourDiscoveryService()) {
-        self.service = service
-    }
-
-    func start() {
-        isBrowsing = true
-        service.start()
-    }
-
-    func refresh() {
-        isBrowsing = true
-        service.refresh()
-    }
-
-    func stop() {
-        isBrowsing = false
-        service.stop()
-    }
-}
-#endif
